3_Multi_Agent_Hierarchical_NATS/WriterAgent/requirements.txt
openai
langgraph
langchain-core
langchain_openai
pydantic
tenacity
nats-py
python-dotenv
